// src/app/open-api/model/wartung.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Wartung { 
    id?: number;
    kundenid?: number;
    vertrag?: number;
    first_quartal?: number;
    second_quartal?: number;
    third_quartal?: number;
    fourth_quartal?: number;
}